# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Finante API
  description: Personal finance management API for tracking transactions, labels, and rules
  version: 1.0.0
  contact:
    name: Finante API Support
    email: cosmin@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.finante.example.com/api/v1
    description: Production server
...
paths:
...
  /transactions/import:
    post:
      tags:
        - transactions
      summary: Import transactions
      description: Import transactions from uploaded files (Excel, CSV)
      operationId: importTransactions
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Files to upload (.xlsx, .xls, .csv)
      responses:
        '200':
          description: Files imported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Files imported successfully"
                  processedFiles:
                    type: integer
                    example: 3
        '400':
          description: Bad request - no files uploaded, unsupported file type, or file too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /rules:
    get:
      tags:
        - rules
      summary: Get all rules
      description: Retrieve all available rules
      responses:
        '200':
          description: Successfully retrieved rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rules'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - rules
      summary: Create rule
      description: Create a new rule
      operationId: createRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleForm'
      responses:
        '201':
          description: Rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleForm'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rules/{id}:
    put:
      tags:
        - rules
      summary: Update rule
      description: Update an existing rule
      operationId: updateRule
      parameters:
        - name: id
          in: path
          required: true
          description: Rule name/ID
          schema:
            type: string
            maxLength: 20
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleForm'
      responses:
        '200':
          description: Rule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '400':
          description: Bad request - validation error or rule name too long
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - rules
      summary: Delete rule
      description: Delete a rule by name/ID
      operationId: deleteRule
      parameters:
        - name: id
          in: path
          required: true
          description: Rule name/ID
          schema:
            type: string
            maxLength: 20
      responses:
        '204':
          description: Rule deleted successfully
        '400':
          description: Bad request - rule name too long
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rules/process:
    post:
      tags:
        - rules
      summary: Apply all rules
      description: Apply all rules to all transactions
      operationId: processRules
      responses:
        '200':
          description: Rules applied successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully applied all rules to all transactions"
                  success:
                    type: boolean
                    example: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Failed to apply rules to transactions"
                  success:
                    type: boolean
                    example: false
  /rules/{id}/process:
    post:
      tags:
        - rules
      summary: Apply specific rule
      description: Apply a specific rule to all transactions
      operationId: processRule
      parameters:
        - name: id
          in: path
          required: true
          description: Rule name/ID
          schema:
            type: string
            maxLength: 20
      responses:
        '200':
          description: Rule applied successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully applied rule to all transactions"
                  success:
                    type: boolean
                    example: true
        '400':
          description: Bad request - rule name too long or empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Rule not found"
                  success:
                    type: boolean
                    example: false
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to get rule"
                  success:
                    type: boolean
                    example: false
components:
  schemas:
...
    Label:
      type: object
      properties:
        href:
          type: string
          format: uri
          example: "/api/v1/labels/123"
        key:
          type: string
          example: "category"
        value:
          type: string
          example: "groceries"
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
          description: Rules that can create this label
...
    RuleForm:
      type: object
      required:
        - name
        - pattern
        - labels
      properties:
        name:
          type: string
          maxLength: 20
          description: Rule name (max 20 characters)
          example: "grocery-rule"
        pattern:
          type: string
          description: Regular expression pattern to match transaction content
          example: "(?i).*(grocery|supermarket).*"
        labels:
          type: object
          additionalProperties:
            type: string
          description: Map of labels to apply when pattern matches
          example:
            category: "groceries"
            type: "expense"
    Rule:
      type: object
      properties:
        href:
          type: string
          format: uri
          example: "/api/v1/rules/grocery-rule"
        name:
          type: string
          example: "grocery-rule"
        pattern:
          type: string
          example: "(?i).*(grocery|supermarket).*"
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
          description: Labels that this rule applies
    Rules:
      type: object
      properties:
        total:
          type: integer
          example: 10
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Validation failed"