# api/v1/types.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package v1

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for TransactionFormKind.
const (
	Credit TransactionFormKind = "credit"
	Debit  TransactionFormKind = "debit"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Error Error message
	Error *string `json:"error,omitempty"`
}

// Label defines model for Label.
type Label struct {
	Href *string `json:"href,omitempty"`
	Key  *string `json:"key,omitempty"`

	// Rules Rules that can create this label
	Rules *[]Rule `json:"rules,omitempty"`
	Value *string `json:"value,omitempty"`
}

// LabelForm defines model for LabelForm.
type LabelForm struct {
	// Key Label key
	Key string `json:"key"`

	// Value Label value
	Value string `json:"value"`
}

// Labels defines model for Labels.
type Labels struct {
	Labels *[]Label `json:"labels,omitempty"`
	Total  *int     `json:"total,omitempty"`
}

// Rule defines model for Rule.
type Rule struct {
	Href *string `json:"href,omitempty"`

	// Labels Labels that this rule applies
	Labels  *[]Label `json:"labels,omitempty"`
	Name    *string  `json:"name,omitempty"`
	Pattern *string  `json:"pattern,omitempty"`
}

// RuleForm defines model for RuleForm.
type RuleForm struct {
	// Labels Map of labels to apply when pattern matches
	Labels map[string]string `json:"labels"`

	// Name Rule name (max 20 characters)
	Name string `json:"name"`

	// Pattern Regular expression pattern to match transaction content
	Pattern string `json:"pattern"`
}

// Rules defines model for Rules.
type Rules struct {
	Rules *[]Rule `json:"rules,omitempty"`
	Total *int    `json:"total,omitempty"`
}

// Transaction defines model for Transaction.
type Transaction struct {
	Account     *int64                         `json:"account,omitempty"`
	Amount      *float32                       `json:"amount,omitempty"`
	Date        *time.Time                     `json:"date,omitempty"`
	Description *string                        `json:"description,omitempty"`
	Href        *string                        `json:"href,omitempty"`
	Kind        *string                        `json:"kind,omitempty"`
	Labels      *[]TransactionLabelAssociation `json:"labels,omitempty"`
}

// TransactionForm defines model for TransactionForm.
type TransactionForm struct {
	// Account Account number
	Account int64 `json:"account"`

	// Amount Transaction amount
	Amount float32 `json:"amount"`

	// Content Transaction description/content
	Content string `json:"content"`

	// Date Transaction date in DD/MM/YYYY format
	Date openapi_types.Date `json:"date"`

	// Kind Transaction type
	Kind TransactionFormKind `json:"kind"`
}

// TransactionFormKind Transaction type
type TransactionFormKind string

// TransactionLabelAssociation defines model for TransactionLabelAssociation.
type TransactionLabelAssociation struct {
	Href *string `json:"href,omitempty"`
	Key  *string `json:"key,omitempty"`

	// RuleHref Reference to the rule that created this label association
	RuleHref *string `json:"rule_href,omitempty"`
	Value    *string `json:"value,omitempty"`
}

// Transactions defines model for Transactions.
type Transactions struct {
	End   *openapi_types.Date `json:"end,omitempty"`
	Items *[]Transaction      `json:"items,omitempty"`
	Start *openapi_types.Date `json:"start,omitempty"`
	Total *int                `json:"total,omitempty"`
}

// GetTransactionsParams defines parameters for GetTransactions.
type GetTransactionsParams struct {
	// StartDate Start date as milliseconds since epoch
	StartDate *int64 `form:"startDate,omitempty" json:"startDate,omitempty"`

	// EndDate End date as milliseconds since epoch
	EndDate *int64 `form:"endDate,omitempty" json:"endDate,omitempty"`
}

// ImportTransactionsMultipartBody defines parameters for ImportTransactions.
type ImportTransactionsMultipartBody struct {
	// Files Files to upload (.xlsx, .xls, .csv)
	Files *[]openapi_types.File `json:"files,omitempty"`
}

// CreateRuleJSONRequestBody defines body for CreateRule for application/json ContentType.
type CreateRuleJSONRequestBody = RuleForm

// UpdateRuleJSONRequestBody defines body for UpdateRule for application/json ContentType.
type UpdateRuleJSONRequestBody = RuleForm

// CreateTransactionJSONRequestBody defines body for CreateTransaction for application/json ContentType.
type CreateTransactionJSONRequestBody = TransactionForm

// ImportTransactionsMultipartRequestBody defines body for ImportTransactions for multipart/form-data ContentType.
type ImportTransactionsMultipartRequestBody ImportTransactionsMultipartBody

// UpdateTransactionJSONRequestBody defines body for UpdateTransaction for application/json ContentType.
type UpdateTransactionJSONRequestBody = TransactionForm

// AddTransactionLabelJSONRequestBody defines body for AddTransactionLabel for application/json ContentType.
type AddTransactionLabelJSONRequestBody = LabelForm